pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cuallee"
version = "1.0.0"

authors = [
  { name="Herminio Vazquez", email="canimus@gmail.com"},
  { name="Virginie Grosboillot", email="vestalisvirginis@gmail.com" }
]
license = {file = "LICENSE"}
description = "Python library for data validation on DataFrame APIs including Snowflake/Snowpark, Apache/PySpark and Pandas/DataFrame."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "toolz >= 0.12.1",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
  "black==24.8.0",
  "ruff==0.7.4",
  "pre-commit==4.0.1",
  "rich == 13.9.4",
]
pyspark = [
  "pyspark>=3.4.0"
]
pyspark_connect = [
  "pyspark[connect]"
]
snowpark = [
  "snowflake-snowpark-python==1.11.1",
  "pyarrow >= 14.0.2"
]
pandas = [
  "pandas>=2.0.1"
]
bigquery =[
  "google-cloud-bigquery>=3.10.0",
  "pyarrow >= 11.0.0"
]
duckdb = [
  "duckdb==1.1.1"
]
polars = [
  "polars>=1.8.2"
]
test = [
  "pytest",
  "pytest-cov",
  "pendulum >= 2.1.2"
]
dagster = [
  "dagster == 1.8.9"
]
cloud = [
  "msgpack == 1.1.0",
]
pdf = [
  "fpdf2==2.7.9"
]
daft = [
  "getdaft == 0.3.2"
]
all = [
  "cuallee[dev,pyspark,pyspark_connect,snowpark,pandas,bigquery,duckdb,polars,test,dagster,cloud,pdf,daft]"
]

[tool.ruff.lint]
# Skip unused variable rules (`F841`).
ignore = ["E731"]


[project.urls]
"Homepage" = "https://github.com/canimus/cuallee"
"Bug Tracker" = "https://github.com/canimus/cuallee"
